Refactor reporting interface with CSP
