+++ atst/domain/csp/__init__.py
@@ -1,11 +1,12 @@
 from .files import RackspaceFileProvider
+from .reports import MockReportingProvider
 
 
 class MockCSP:
-
-    def __init__(self, file_provider):
-        self.files = file_provider
+    def __init__(self, app):
+        self.files = RackspaceFileProvider(app)
+        self.reports = MockReportingProvider()
 
 
 def make_csp_provider(app):
-    app.csp = MockCSP(RackspaceFileProvider(app))
+    app.csp = MockCSP(app)
